week8_basic_streamlit/requirements.txt
streamlit
snowflake-connector-python
snowflake-snowpark-python
snowflake-ml-python[all]
pandas
